+++ README.md
@@ -2,3 +2,4 @@
 Algorithms and Data Structures
 
 
+* [Union Find](src/main/java/com/rehmanz/dynamic_connectivity/README.md) is used to to dynamically maintain information about connected components of a graph
